Create calculator using Calculation class

+++ src/main/java/List.java
@@ -1,12 +1,12 @@
 public class List {
     final int MAX_SIZE = 5;
-    int[] array = new int[MAX_SIZE];
+    double[] array = new double[MAX_SIZE];
     int size;
 
-    private int index(int element){
+    private int index(double element) {
 
-        for(int i = 0; i < size; i++) {
-            if (array[i] == element){
+        for (int i = 0; i < size; i++) {
+            if (array[i] == element) {
                 return i;
             }
         }
@@ -18,16 +18,16 @@ public class List {
         return size;
     }
 
-    public void add(int element) {
+    public void add(double element) {
         array[size++] = element;
     }
 
-    public void remove(int element) {
+    public void remove(double element) {
         int indexOfElement = index(element);
         array[indexOfElement] = array[size--];
     }
 
-    public int element(int index) {
+    public double element(int index) {
         return array[index];
     }
